requirements.txt
aioredis==2.0.1
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.0.1
black==24.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.0
fastapi-mail==1.4.1
fastapi-redis-cache==0.2.5
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
